change to typescript
in order to make it work on android
